refactor: Separate ContactViewModel into its own file
- Added coding standard: Each class must be in its own file
- Moved ContactViewModel from ContactsListViewModel.cs to ContactViewModel.cs
- Follows the one-class-per-file principle for better code organization
- No functional changes, just file restructuring
