+++ src/UnoApp/UnoApp/Presentation/Pages/Contacts/ContactViewModel.cs
@@ -1,0 +1,138 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+using Microsoft.Extensions.Logging;
+using Microsoft.UI.Xaml;
+using Shiny.Mediator;
+using Uno.Extensions.Navigation;
+using UnoApp.ApiClient;
+using DtoModels = SharedModels.Dtos.Contacts;
+
+namespace UnoApp.Presentation.Pages.Contacts
+{
+    public partial class ContactViewModel : ObservableObject
+    {
+        private readonly UnoApp.ApiClient.ContactDto _contact;
+        private readonly IMediator _mediator;
+        private readonly INavigator _navigator;
+        private readonly int _defaultCallbackDays;
+        private readonly ILogger _logger;
+
+        public ContactViewModel(UnoApp.ApiClient.ContactDto contact, IMediator mediator, INavigator navigator, int defaultCallbackDays)
+        {
+            _contact = contact;
+            _mediator = mediator;
+            _navigator = navigator;
+            _defaultCallbackDays = defaultCallbackDays;
+            _logger = LoggerFactory.Create(builder => builder.AddConsole()).CreateLogger<ContactViewModel>();
+        }
+
+        public string Name => _contact.Name;
+        public string Email => _contact.Email;
+        public string Phone => _contact.Phone;
+        public string Branche => _contact.Industry;
+        public string? CallNotes => _contact.CallNotes;
+        public DateTime? NextCallDate => _contact.NextCallDate?.DateTime;
+        
+        public string NextCallDateDisplay
+        {
+            get
+            {
+                if (NextCallDate == null)
+                    return "Nicht geplant";
+                    
+                var today = DateTime.Today;
+                var callDate = NextCallDate.Value.Date;
+                
+                if (callDate == today)
+                    return "Heute";
+                else if (callDate == today.AddDays(1))
+                    return "Morgen";
+                else if (callDate < today)
+                    return $"Überfällig ({callDate:dd.MM.})";
+                else
+                    return callDate.ToString("dd.MM.yyyy");
+            }
+        }
+        
+        public string NextCallDateColor
+        {
+            get
+            {
+                if (NextCallDate == null)
+                    return "Gray";
+                    
+                var today = DateTime.Today;
+                if (NextCallDate.Value.Date < today)
+                    return "Red";
+                else if (NextCallDate.Value.Date == today)
+                    return "Orange";
+                else
+                    return "Black";
+            }
+        }
+        
+        public string StatusText => ((DtoModels.CallStatus)(int)_contact.CallStatus) switch
+        {
+            DtoModels.CallStatus.New => "Neu",
+            DtoModels.CallStatus.Scheduled => "Geplant",
+            DtoModels.CallStatus.Reached => "Erreicht",
+            DtoModels.CallStatus.NotReached => "Nicht erreicht",
+            DtoModels.CallStatus.Completed => "Abgeschlossen",
+            DtoModels.CallStatus.Postponed => "Verschoben",
+            _ => "Unbekannt"
+        };
+        
+        public string StatusColor => ((DtoModels.CallStatus)(int)_contact.CallStatus) switch
+        {
+            DtoModels.CallStatus.New => "Blue",
+            DtoModels.CallStatus.Scheduled => "Orange",
+            DtoModels.CallStatus.Reached => "Green",
+            DtoModels.CallStatus.NotReached => "Red",
+            DtoModels.CallStatus.Completed => "Gray",
+            DtoModels.CallStatus.Postponed => "Purple",
+            _ => "Gray"
+        };
+        
+        public Visibility HasNotes => string.IsNullOrWhiteSpace(CallNotes) ? Visibility.Collapsed : Visibility.Visible;
+
+        [RelayCommand]
+        private async Task CallAsync()
+        {
+            // Navigate to call detail page
+            await _navigator.NavigateDataAsync(
+                this,
+                data: new Dictionary<string, object> { ["Contact"] = _contact },
+                cancellation: CancellationToken.None
+            );
+        }
+
+        [RelayCommand]
+        private async Task RescheduleAsync()
+        {
+            // TODO: Show date picker dialog
+            var newDate = DateTime.Today.AddDays(_defaultCallbackDays);
+            
+            // TODO: RescheduleCallHttpRequest needs to be generated from WebApi.json
+            // For now, we'll just update the local state
+            _logger.LogInformation("Would reschedule contact {ContactId} to {Date}", _contact.Id, newDate);
+        }
+
+        [RelayCommand]
+        private async Task MarkReachedAsync()
+        {
+            // TODO: UpdateCallStatusHttpRequest needs to be generated from WebApi.json
+            _logger.LogInformation("Would mark contact {ContactId} as reached", _contact.Id);
+        }
+
+        [RelayCommand]
+        private async Task MarkNotReachedAsync()
+        {
+            // TODO: UpdateCallStatusHttpRequest needs to be generated from WebApi.json
+            _logger.LogInformation("Would mark contact {ContactId} as not reached", _contact.Id);
+        }
+    }
+}
